requirements.txt
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
matplotlib==3.7.1
requests==2.31.0
streamlit==1.34.0
streamlit-cropper==0.2.2
opencv-contrib-python==4.9.0.80     
torch==2.2.2
torchvision==0.17.2
faiss-cpu==1.8.0
tqdm==4.66.4             
scikit-image==0.24.0
